+++ CHANGES.rst
@@ -29,7 +29,8 @@ Bug fixes:
 - Modernize robot keywords that use "Get Element Attribute" [ale-rt] (#2615)
 - Fix metabundle resource ordering to pay attention to depends setting
   [vangheem] (#2641)
-
+- Do not save type settings in "content-controlpanel" when switching between types.
+  [cekk]
 
 Changelog
 =========